+++ benchmarks/storm-bench/src/main/java/fr/unice/yamb/storm/bolts/WindowedBusyWaitBolt.java
@@ -18,7 +18,7 @@ public class WindowedBusyWaitBolt extends BaseWindowedBolt {
     private String _me;
 
 
-    public WindowedBusyWaitBolt(long cycles, float frequency){
+    public WindowedBusyWaitBolt(long cycles, double frequency){
         this._cycles = cycles;
         if (frequency > 0) this._rate = (int)(1/frequency);
         else this._rate = 0;
